requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
luno-python==0.0.7
python-decouple==3.4
requests==2.25.1
six==1.15.0
urllib3==1.26.3
